API/requirements.txt
#pip install -r requirements.txt 

fastapi
uvicorn[standard]
pyodbc
python-jose
passlib[argon2]
pydantic
python-multipart

